* sidebar: fix read from freed pointer after ARRAY_REMOVE
 * lua: fix resource leak on error paths
 * conn/raw: fix truncation of write() return value
 * conn/tunnel: fix truncation of write() return value
 * main: explicitly ignore mutt_autocrypt_init() return value
 * send: fix type mismatch in message-id generation
 * pager: explicitly ignore mutt_file_seek() return in cleanup path

+++ send/expando_msgid.c
@@ -275,7 +275,7 @@ char *msgid_generate(void)
   if (buf_at(buf, 0) != '<')
     buf_insert(buf, 0, "<");
 
-  const int last = buf_len(buf) - 1;
+  const size_t last = buf_len(buf) - 1;
   if (buf_at(buf, last) != '>')
     buf_addch(buf, '>');
